* Fixing bug in FASTA reading   -- The bug was related with some sequence identifiers containing the `>` character in the middle. Our parser was not expecting that was considering that as the start of a new sequence.

+++ src/ParallelFastaReader.cpp
@@ -10,7 +10,7 @@
 #include "../include/DistributedFastaData.h"
 
 
-void ParallelFastaReader::read_fasta(const char *file, ushort overlap, int rank,
+void ParallelFastaReader::read_fasta(const char *file, uint64_t overlap, int rank,
                                      int world_size, char *&buff,
                                      uint64_t &l_start, uint64_t &l_end) {
